Speedup

+++ service/build.py
@@ -1,68 +1,52 @@
 import os
+from service.sync import SyncService
 
-class BuildSync(object):
-    def __init__(self, source_rack, dest_rack, logger):
-        self.source_rack = source_rack
-        self.dest_rack   = dest_rack
-        self.logger      = logger
+class BuildSync(SyncService):
+    
+    def sync(self, app):
+        self.app_name = app['name']
 
-    def sync(self, source_apps=None):
-        source_apps = source_apps if source_apps else self.source_rack.apps.get()
+        self._sync(
+            self._compare(app)
+        )
 
-        requiring_update = self._compare(source_apps)
-
-        self._sync(requiring_update)
-
-    def _sync(self, apps):
+    def _sync(self, app):
         """ Export build artifact from source app and import it to the destination app """
+        if not app:
+            return
 
-        for app in apps:
-            app_name = app['name']
+        self._log('Syncing build')
 
-            self.logger.info('Syncing build for {}.{}'.format(self.dest_rack.name(), app_name))
+        active_build_id = self.source_rack.app(self.app_name).builds.active_build_id()
+        tmp_file        = '/tmp/{}.tgz'.format(active_build_id)
 
-            active_build_id = self.source_rack.app(app_name).builds.active_build_id()
-            tmp_file        = '/tmp/{}.tgz'.format(active_build_id)
+        self.source_rack.app(self.app_name).builds.export_build(active_build_id, tmp_file)
+        self.dest_rack.app(self.app_name).builds.import_build(active_build_id, tmp_file)
+        os.remove(tmp_file)
 
-            self.source_rack.app(app_name).builds.export_build(active_build_id, tmp_file)
-            self.dest_rack.app(app_name).builds.import_build(active_build_id, tmp_file)
-            os.remove(tmp_file)
+        return app
 
-        return apps
-
-    def _compare(self, apps):
+    def _compare(self, app):
         """ Compare source and destination apps to see if the build id are the same or not """
 
-        self.logger.info('Comparing builds')
+        if not app:
+            return
 
-        mismatched_build = []
+        self._log('Comparing build')
 
-        for app in apps:
-            app_name = app['name']
+        source_build_id = self.source_rack.app(self.app_name).builds.active_build_id()
 
-            source_build_id = self.source_rack.app(app_name).builds.active_build_id()
+        if not source_build_id:
+            return None
 
-            if not source_build_id:
-                continue
+        #check if this build already exists on the dest rack
+        dest_build = self.dest_rack.app(self.app_name).builds.get(source_build_id)
 
-            #check if this build already exists on the dest rack
-            dest_build = self.dest_rack.app(app_name).builds.get(source_build_id)
+        if dest_build and 'error' not in dest_build:
+            self._log('Build {} exists'.format(source_build_id))
 
-            if dest_build and 'error' not in dest_build:
-                self.logger.info('Build {} exists on {}.{}'.format(
-                    source_build_id,
-                    self.dest_rack.name(),
-                    app_name
-                ))
+            return None
 
-                continue
+        self._log('Build {} not found'.format(source_build_id))
 
-            self.logger.info('Build {} not found on {}.{}'.format(
-                source_build_id,
-                self.dest_rack.name(),
-                app_name
-            ))
-
-            mismatched_build.append(app)
-
-        return mismatched_build
+        return app
